Refactor game existence check in GameController and GameService

+++ src/main/java/CENG453/group18/repository/GameRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import CENG453.group18.entity.Player;
 
-public interface GameRepository extends JpaRepository<Game, Long> {
-    Game getGameByGameID(int id);
-    void deleteGameByGameID(int id);
-    boolean existsGameByGameID(int id);
+public interface GameRepository extends JpaRepository<Game, Integer> {
+    Game getGameByGameID(Integer id);
+    void deleteGameByGameID(Integer id);
+    boolean existsGameByGameID(Integer id);
     Game findByPlayer1OrPlayer2OrPlayer3OrPlayer4(Player player1, Player player2, Player player3, Player player4);
 }
